feat: add draft Members section on Lobby Page
