+++ client/src/components/pages/lobby-page/game-settings-section/game-settings-section.sass
@@ -1,12 +1,4 @@
 @import '@styles/variables'
-.lobby-page
-  margin: 0 auto
-  display: flex
-  flex-direction: column
-  align-items: center
-  justify-content: center
-  margin-bottom: 50px
-
 .welcome-header
   font-size: 48px
   font-weight: 700
@@ -24,9 +16,6 @@
 
 .input-lobby-container
   margin-bottom: 30px
-
-.section-header
-  text-align: center
 
 .switch-lobby
   display: flex
