+++ client/src/components/pages/lobby-page/lobby-page.sass
@@ -6,3 +6,7 @@
   align-items: center
   justify-content: center
   margin-bottom: 50px
+
+  .section-header
+    text-align: center
+    margin: 50px 0
