+++ client/src/components/pages/lobby-page/lobby-page.tsx
@@ -1,13 +1,13 @@
-import GameSettingsSection from '@components/game-settings-section/game-settings-section';
+import GameSettingsSection from '@components/pages/lobby-page/game-settings-section/game-settings-section';
+import MembersSection from '@components/pages/lobby-page/members-section/members-section';
 import './lobby-page.sass';
 
 const LobbyPage = () => (
   <div className="container">
     <div className="lobby-page">
       <div className="lobby-content">
-        <section className="lobby-content-game-settings">
-          <GameSettingsSection />
-        </section>
+        <MembersSection />
+        <GameSettingsSection />
       </div>
     </div>
   </div>
