+++ client/src/components/pages/lobby-page/members-section/members-section.sass
@@ -1,0 +1,5 @@
+.members-lobby
+  display: grid
+  grid-template-columns: repeat(3, 1fr)
+  grid-template-rows: auto
+  gap: 15px
