+++ client/src/components/ui/memberCard/memberCard.sass
@@ -1,0 +1,19 @@
+
+.member-card
+  width: 280px
+  height: 76px
+  padding: 0 10px 0 10px
+  display: flex
+  column-gap: 10px
+  align-items: center
+
+  &-delete-btn
+    width: 30px
+    height: 30px
+    background-image: url('../../../assets/icons/deleteMember.svg')
+
+  &-header
+    width: 200px
+    overflow: hidden
+    white-space: nowrap
+    padding: 0
