+++ client/src/models/types.ts
@@ -16,3 +16,37 @@ export interface IInputButtonProps {
    */
   valueHandler: (value: string) => void;
 }
+
+export interface ICustomAvatar {
+  /*
+   * first name of User
+   */
+  firstName: string;
+  /*
+   * last name of User
+   */
+  lastName?: string;
+  /*
+   * avatar image
+   */
+  avatarImage?: string;
+}
+
+export interface IMemberCard {
+  /*
+   * first name of User
+   */
+  firstName: string;
+  /*
+   * last name of User
+   */
+  lastName?: string;
+  /*
+   * role in the team
+   */
+  role: string;
+  /*
+   * avatar image
+   */
+  avatarImage?: string;
+}
